fix: correct int/int64 type mismatch in body limit middleware build break
NewBodyLimitMiddleware (added in #157) compared c.Request().Header.ContentLength()
(int) directly against int64(limit), which fails to compile ("mismatched types
int and int64"). This broke go build on main immediately after the merge.
Convert contentLength to int64 before comparing.

+++ internal/api/bodylimit.go
@@ -34,7 +34,7 @@ func NewBodyLimitMiddleware(cfg BodyLimitConfig) fiber.Handler {
 		}
 
 		contentLength := c.Request().Header.ContentLength()
-		if contentLength > int64(limit) {
+		if int64(contentLength) > int64(limit) {
 			return fiber.NewError(fiber.StatusRequestEntityTooLarge, "Request body exceeds size limit")
 		}
